Split resolver replay validation tests
